web/requirements.txt
Flask==3.1.0
Jinja2==3.1.6
PyMySQL==1.1.1
python-dotenv==1.1.0